+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `ContentCard`: Add component.
+- Feature: `TextButton`: Add component.
+- Feature: `AirplaneIconNext`: Add component.
+- Feature: `BubbleIconNext`: Add component.
+
 ## [6.6.0] - 2021-12-02
 
 Features